refactor: move BottomNavbar to common components for better organization
