+++ src/components/ActiveReservationPage/index.tsx
@@ -1,41 +1,52 @@
+import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import BottomNavBar from "@/components/BottomNavbar";
+import BottomNavBar from "@/components/common/BottomNavbar";
 import { useActiveReservation } from "@/hooks/useActiveReservation";
-import { PageHeader } from "./PageHeader";
+import { Header } from "@/components/common/Header";
 import { ReservationInfoCard } from "./ReservationInfoCard";
 import { ServicesCard } from "./ServicesCard";
 import { ProductsCard } from "./ProductsCard";
 import { HomeButton } from "./HomeButton";
+import PolicyNotice from "@/components/ActiveReservationPage/PolicyNotice";
 
-function ActiveReservationPage() {
+export default function ActiveReservationPage() {
   const { t } = useTranslation();
-  const { activeReservation, handleReservationAction, getActionButtonText } = useActiveReservation();
+  const [showPolicyNotice, setShowPolicyNotice] = useState(true);
+  const { activeReservation, handleReservationAction, getActionButtonText } =
+    useActiveReservation();
 
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setShowPolicyNotice(false)
+    }, 5000)
+    return () => clearTimeout(timer)
+  }, [])
+
+  const title = t("activeReservation.title");
+  
   if (!activeReservation) {
     return <div>{t("activeReservation.notFound")}</div>;
   }
+  
+  const isReservationEnded = new Date() > new Date(activeReservation.endTime);
 
   return (
     <div className="min-h-screen bg-background text-foreground pb-16">
-      <PageHeader />
-      
+      <Header title={title} showBackButton={true} />
       <div className="p-4">
-        <ReservationInfoCard 
-          reservation={activeReservation} 
+        {!isReservationEnded && showPolicyNotice && (
+          <PolicyNotice />
+        )}
+        <ReservationInfoCard
+          reservation={activeReservation}
           onAction={handleReservationAction}
           actionButtonText={getActionButtonText(t)}
         />
-        
-        <ServicesCard services={activeReservation.services} />
-        
-        <ProductsCard products={activeReservation.products} />
-        
+        <ServicesCard services={activeReservation.services ?? []} />
+        <ProductsCard products={activeReservation.products ?? []} />
         <HomeButton />
       </div>
-      
       <BottomNavBar />
     </div>
   );
 }
-
-export default ActiveReservationPage;
