+++ src/components/FoodPage.tsx
@@ -2,7 +2,7 @@ import { ArrowLeft, Plus, Minus } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Card } from "@/components/ui/card";
 import { Separator } from "@/components/ui/separator";
-import BottomNavBar from "./BottomNavbar";
+import BottomNavBar from "./common/BottomNavbar";
 import { useNavigate } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { useEffect, useState } from "react";
