+++ src/components/FoodPage/index.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react"
 import { useTranslation } from "react-i18next"
-import BottomNavBar from "@/components/BottomNavbar"
+import BottomNavBar from "@/components/common/BottomNavbar"
 import { useNavigate } from "react-router-dom"
 import { useTheme } from "@/context/ThemeContext"
-import { Header } from "@/components/FoodPage/Header"
+import { Header } from "@/components/common/Header";
 import { PolicyNotice } from "@/components/FoodPage/PolicyNotice"
 import { ProductCard } from "@/components/FoodPage/ProductCard"
 import { ConfirmButton } from "@/components/FoodPage/ConfirmButton"
@@ -40,7 +40,7 @@ export default function FoodPage() {
 
   return (
     <div className={`min-h-screen bg-background text-foreground pb-16 ${theme}`}>
-      <Header />
+      <Header title={t("food.title")} showBackButton={true} />
       {showPolicyNotice && <PolicyNotice showButton={showButton} />}
       <div className="px-6">
         {Object.entries(productData).map(([category, products]) => (
